make lease recovery comply with the contract
